requirements.txt
fastapi[standard]
uvicorn
pydantic
sqlalchemy
pymysql
mysql
python-dotenv
python-multipart
aiosmtplib
python-jose
pydantic[email]
requests
matplotlib
openpyxl
dataclasses
bcrypt